Decouple seq2seq and postnet
so that we can train seq2seq2 and postnet separately

also changed the converter inputs to decoder outputs, instead of decoder
states. Comments added
