+++ deepvoice3_pytorch/nyanko.py
@@ -7,24 +7,7 @@ from torch.autograd import Variable
 import math
 import numpy as np
 
-from fairseq.models.fconv import Embedding, Linear
-
-
-def Conv1d(in_channels, out_channels, kernel_size, dropout=0, **kwargs):
-    from .conv import Conv1d
-    m = Conv1d(in_channels, out_channels, kernel_size, **kwargs)
-    std = math.sqrt((4 * (1.0 - dropout)) / (m.kernel_size[0] * in_channels))
-    m.weight.data.normal_(mean=0, std=std)
-    m.bias.data.zero_()
-    return nn.utils.weight_norm(m)
-
-
-def ConvTranspose1d(in_channels, out_channels, kernel_size, dropout=0, **kwargs):
-    m = nn.ConvTranspose1d(in_channels, out_channels, kernel_size, **kwargs)
-    std = math.sqrt((4 * (1.0 - dropout)) / (m.kernel_size[0] * in_channels))
-    m.weight.data.normal_(mean=0, std=std)
-    m.bias.data.zero_()
-    return nn.utils.weight_norm(m)
+from .module import Embedding, Linear, Conv1d, ConvTranspose1d
 
 
 class Converter(nn.Module):
